pyproject.toml
[project]
name = "openmedaxis"
version = "0.0.1"
description = "Open framework for generative medical imaging research"
authors = [{name="AllAxisAI"}]
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "scipy",
    "pytorch-lightning",
    "nibabel",
    "tqdm"
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["oma*"]
exclude = ["assets*", "configs*", "tests*"]